Refactored the contracts, tests and added missing require() messages
