+++ contracts/Migrations.sol
@@ -1,4 +1,6 @@
-pragma solidity ^0.6.6;
+// SPDX-License-Identifier: MIT
+
+pragma solidity ^0.6.8;
 
 contract Migrations {
   address public owner;
