+++ contracts/token/ERC20/ERC20.sol
@@ -25,13 +25,16 @@ TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
 SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 
-pragma solidity ^0.6.6;
+// SPDX-License-Identifier: MIT
+
+pragma solidity ^0.6.8;
 
 import "@openzeppelin/contracts/GSN/Context.sol";
 import "@openzeppelin/contracts/introspection/IERC165.sol";
 import "@openzeppelin/contracts/math/SafeMath.sol";
 import "@openzeppelin/contracts/utils/Address.sol";
 import "./IERC20.sol";
+import "./IERC20Detailed.sol";
 
 /**
  * @dev Implementation of the {IERC20} interface.
@@ -57,15 +60,16 @@ import "./IERC20.sol";
  * functions have been added to mitigate the well-known issues around setting
  * allowances. See {IERC20-approve}.
  */
-contract ERC20 is Context, IERC165, IERC20 {
+abstract contract ERC20 is Context, IERC165, IERC20, IERC20Detailed {
+
     using SafeMath for uint256;
     using Address for address;
 
-    mapping (address => uint256) private _balances;
+    mapping (address => uint256) internal _balances;
 
-    mapping (address => mapping (address => uint256)) private _allowances;
+    mapping (address => mapping (address => uint256)) internal _allowances;
 
-    uint256 private _totalSupply;
+    uint256 internal _totalSupply;
 
     /**
      * @dev Check if support an interface id
@@ -77,7 +81,15 @@ contract ERC20 is Context, IERC165, IERC20 {
             // ERC165 interface id
             interfaceId == 0x01ffc9a7 ||
             // ERC20 interface id
-            interfaceId == 0x36372b07
+            interfaceId == 0x36372b07 ||
+            // ERC20Name interface id
+            interfaceId == 0x06fdde03 ||
+            // ERC20Symbol interface id
+            interfaceId == 0x95d89b41 ||
+            // ERC20Decimals interface id
+            interfaceId == 0x313ce567 ||
+            // ERC20Detailed interface id
+            interfaceId == 0xa219a025
         );
     }
