+++ test/contracts/token/ERC20/behaviors/ERC20WithOperators.behavior.js
@@ -1,8 +1,8 @@
 const { BN, constants, expectEvent, expectRevert } = require('@openzeppelin/test-helpers');
 const { MAX_UINT256, ZERO_ADDRESS } = constants;
 
-function shouldBehaveLikeERC20FullWhitelistedOperators(initialSupply, initialHolder, recipient, [ spender ]) {
-    describe('ERC20 Base is WhitelistedOperators', function () {
+function shouldBehaveLikeERC20WithOperators(initialSupply, initialHolder, recipient, [ spender ]) {
+    describe('like an ERC20WithOperators', function () {
         beforeEach(async function () {
             await this.token.whitelistOperator(spender, true, {from: initialHolder});
         });
@@ -403,5 +403,5 @@ function shouldBehaveLikeERC20FullWhitelistedOperators(initialSupply, initialHol
 }
 
 module.exports = {
-    shouldBehaveLikeERC20FullWhitelistedOperators
+    shouldBehaveLikeERC20WithOperators
 };
